refactor(auth):  project architecture
