+++ backend/auth/internal/handlers/auth_session.go
@@ -1,7 +1,7 @@
 package handlers
 
 import (
-	"auth/backend/services/auth/internal/models"
+	"backend/auth/internal/models"
 	"log"
 	"time"
