+++ infra/services/auth/Dockerfile
@@ -1,24 +1,20 @@
-# ==========================================
-# Stage 1: Builder
-# ==========================================
 FROM golang:1.25-alpine AS builder
 
 RUN apk add --no-cache git
 
 WORKDIR /app
 
-# Puisque le contexte est à la racine, go.mod est directement accessible
-COPY go.mod go.sum ./
+COPY ./backend/auth/go.mod ./backend/auth/go.sum ./
+COPY ./backend/shared/go.mod ./backend/shared/go.sum ./
+
 RUN go mod download
 
-# On copie uniquement le dossier de ton service d'authentification
-COPY ./backend ./backend
+COPY ./backend/auth .
+COPY ./backend/shared ../shared
 
-# On compile en pointant vers le bon chemin relatif
-RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -ldflags="-w -s" -o ft_box_auth ./backend/services/auth/cmd
-# ==========================================
-# Stage 2: Runner (Production/Dev ready)
-# ==========================================
+RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -ldflags="-w -s" -o ft_box_auth ./cmd/api
+
+
 FROM alpine:3.19
 
 WORKDIR /app
